% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtdd_get_yield_curve.R
\name{get.yield.curve}
\alias{get.yield.curve}
\title{Gets the current yield curve}
\usage{
get.yield.curve()
}
\value{
A data frame with information about the yield curve.
}
\description{
Downloads and parses information about the current Brazilian yield curve from Anbima.
}
\examples{
\dontrun{
df.yield <- get.yield.curve()
str(df.yield)
}
}
