% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cache_folder}
\alias{get_cache_folder}
\title{Returns the cache directory path}
\usage{
get_cache_folder()
}
\value{
A character string representing the folder path.
}
\description{
Returns the cache directory path
}
\examples{
get_cache_folder()
}
