% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtdd_get_yield_curve.R
\name{get_yield_curve}
\alias{get_yield_curve}
\alias{get.yield.curve}
\title{Gets the current yield curve}
\usage{
get_yield_curve()

get.yield.curve()
}
\value{
A data frame with information about the yield curve.
}
\description{
Downloads and parses information about the current Brazilian yield curve from Anbima.
}
\examples{
\dontrun{
df_yield <- get_yield_curve()
str(df_yield)
}
}
