% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_yield_curve}
\alias{plot_yield_curve}
\title{Plot the ANBIMA Yield Curve}
\usage{
plot_yield_curve(df_yc)
}
\arguments{
\item{df_yc}{A data frame or tibble returned by \code{\link{get_yield_curve}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots the yield curve data frame downloaded using \code{\link{get_yield_curve}}.
}
\examples{
\dontrun{
df_yc <- get_yield_curve()
plot_yield_curve(df_yc)
}
}
