% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_td_names}
\alias{get_td_names}
\title{Returns the available asset names at the Tesouro Direto (TD) website}
\usage{
get_td_names()
}
\value{
A character vector of names.
}
\description{
Returns the available asset names at the Tesouro Direto (TD) website
}
\examples{
get_td_names()
}
