% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_td_names}
\alias{get_td_names}
\title{Returns available nammes at TD site}
\usage{
get_td_names()
}
\value{
string vector
}
\description{
Returns available nammes at TD site
}
\examples{
get_td_names()
}
