% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_td.R
\name{td_get}
\alias{td_get}
\title{Downloads data of Brazilian government bonds directly from the website}
\usage{
td_get(
  asset_codes = "LTN",
  first_year = 2005,
  last_year = lubridate::year(Sys.Date()),
  dl_folder = get_cache_folder()
)
}
\arguments{
\item{asset_codes}{Strings that identify the assets (1 or more assets) in the
names of the excel files. E.g. asset_codes = 'LTN'. When set to NULL, it
will download all available assets}

\item{dl_folder}{Name of folder to save excel files from tesouro direto (will
create if it does not exists)}

\item{do_clean_up}{Clean up folder before downloading? (TRUE or FALSE)}

\item{do_overwrite}{Overwrite excel files? (TRUE or FALSE). If FALSE, will
only download the new data for the current year}

\item{n_dl}{Sets how many files to download from the website. Used only to
decrease CRAN CHECK time. The default value is NULL (downloads all files)}
}
\value{
TRUE if successful
}
\description{
This function looks into the tesouro direto website
(https://www.tesourodireto.com.br/) and
downloads all of the files containing prices and yields of government bonds.
You can use input asset_codes to restrict the downloads to specific bonds
}
\examples{
# only download file where string LTN is found
# (only 1 file for simplicity)
\dontrun{
download.TD.data(asset_codes = 'LTN', n_dl = 1)
}
# The excel file should be available in folder 'TD Files' (default name)

}
