% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_get_current.R
\name{td_get_current}
\alias{td_get_current}
\title{Returns current TD prices}
\usage{
td_get_current()
}
\value{
A data frame with prices.
}
\description{
Fetches current prices of Tesouro Direto (TD) assets from the website's JSON API at <https://www.tesourodireto.com.br/titulos/precos-e-taxas.htm>.
}
\examples{
\dontrun{
df_current <- td_get_current()
}
}
