% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_get_current.R
\name{td_get_current}
\alias{td_get_current}
\title{Returns current TD prices}
\usage{
td_get_current()
}
\value{
a dataframe with prices
}
\description{
Fetches current prices of TD assets from website's json api at <https://www.tesourodireto.com.br/titulos/precos-e-taxas.htm>
}
\examples{
td_get_current()
}
